# Add RISC-V V extension support and add R-V V isal_adler32