--- a/igzip/riscv64/Makefile.am
+++ b/igzip/riscv64/Makefile.am
@@ -29,4 +29,5 @@
 
 lsrc_riscv64 += \
 	igzip/riscv64/igzip_multibinary_riscv64_dispatcher.c \
-	igzip/riscv64/igzip_multibinary_riscv64.S
+	igzip/riscv64/igzip_multibinary_riscv64.S \
+	igzip/riscv64/igzip_isal_adler32_rvv.S