--- a/raid/Makefile.am
+++ b/raid/Makefile.am
@@ -33,6 +33,7 @@ lsrc        += 	raid/raid_base.c
 
 lsrc_base_aliases += raid/raid_base_aliases.c
 lsrc_ppc64le      += raid/raid_base_aliases.c
+lsrc_riscv64      += raid/raid_base_aliases.c
 
 lsrc_x86_64 += \
 		raid/xor_gen_sse.asm \