@hleft 2025-03-03
> Thanks @hleft. Our internal CI is throwing this warning:
> 
> configure.ac:80: warning: AC_COMPILE_IFELSE was called before AC_USE_SYSTEM_EXTENSIONS
> 
> Could you take a look at it and fix it? This is caused by the last commit, I believe.

Okay, I have updated its order.